fix: silence C4244 in analyticalSES.C — explicit double->float cast (v1.6.2)
Site: source/STRUCTURE/analyticalSES.C:116 (inside calculateSESAtomAreas).
The hash map value type is float (HashMap<const Atom*, float> from the
public analyticalSES.h:31 API contract) but tmp_atom_areas[i] is
double-precision (output of the connolly_ Fortran routine that does
the actual SES computation). The implicit narrowing was happening
silently on gcc/clang; MSVC /W4 flags it as C4244 ("possible loss of
data").

Not a correctness bug — the float-precision contract is intentional
(saves 4 bytes/atom storage; SES rounding error is dominated by
surface tessellation anyway, so single-precision is sufficient). The
fix is a one-line static_cast<float>(...) that documents the intent
and silences the warning. Local macOS BALL rebuild green ([100%]
Built target BALL).

Targets v1.6.2 (lands on v1.6-modernization after the v1.6.1 tag).
Same v1.6.2 "small wins, no substrate change" envelope as Phase 999.26
(C4910 suppress) — cosmetic Windows-warning closures that don't fit
Phase 999.22's CENSUS-ONLY policy because they're individual one-site
fixes, not a broad warning-execution pass.

+++ source/STRUCTURE/analyticalSES.C
@@ -113,7 +113,11 @@ namespace BALL
 		atom_areas.clear();
 		for (Position i = 0; i < atoms.size(); ++i)
 		{
-			atom_areas.insert(std::pair<const Atom*, float>(atoms[i], tmp_atom_areas[i]));
+			// Explicit double -> float narrowing. The connolly_ Fortran routine returns
+			// double-precision atom areas; the HashMap<const Atom*, float> API contract
+			// stores float (saves 4 bytes/atom; SES rounding error is dominated by
+			// surface tessellation anyway). Cast documents the intent + silences C4244.
+			atom_areas.insert(std::pair<const Atom*, float>(atoms[i], static_cast<float>(tmp_atom_areas[i])));
 		}
 
 		// free the input fields
